Pin ESM `module` to es6 and `moduleResolution` to node10
I forgot how stupid this dual-build setup is. If you tell typescript to
build `node16` modules, it checks the package.json `type` field to
decide whether you want cjs or esm.

We generate package.json overrides for the cjs and esm builds **after**
compiling with tsc, and we can't just put one or the other in the normal
package.json file. We also seemingly can't feed tsc any sort of override
for the `type` value.

And so, back to the previous approach - locking to es6 instead of esnext
this time, but inspecting the build artifacts tells me that there's no
practical difference.